Removed Redundant Method Call
Made a minor change to the processFacebookPageFeedStatus function: assigned the content of request_until_succeed(url) to a variable for each URL the function constructs, and then referenced that variable in the if statements that determine whether the status actually has content (aka whether it’s been deleted). That keeps the function from being called twice, and therefore requesting the same information about the status twice and unnecessarily doubling the amount of information we ask Facebook for. Since this was in the martha-playingaroundpt2.py file, which has already been run, I have no idea if it’s useful at all, but it was easy and still felt kind of important, so I wanted to commit it anyway in case we need it in a later file.

Also deleted a now-unnecessary series of comments.